Define transfer / deposit/ withdraw responses

+++ src/main/java/com/alshubaily/fintech/tiger_ledger_service/model/account/request/WithdrawRequest.java
@@ -3,7 +3,9 @@ package com.alshubaily.fintech.tiger_ledger_service.model.account.request;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigInteger;
+
 public record WithdrawRequest(
-        @Min(1) long accountId,
+        @Min(1) BigInteger accountId,
         @Positive double amount
 ) {}
